Do not crash if RDKit fails ot parse smiles
There seem to be some cases where RDKit fails to parse a SMILES string
that works with OpenBabel. Not sure what these are, but this fix
prevents parsing failures from throwing exceptions. I would imagine
these changes are small enough to not be a huge issue, but is something
we should figure at as a low priority